Version 0.0.1
Added a Satemachine for debouncing buttons that will be used for
controlling the microcontroller. First idea is to use joystick buttons
up,down,left and right for navigation in the menus and to have some kind
of an confim and cancel button.

So in the end 6 buttons has to be unbounced to make sure only one pressing
of a button is recognized.

Signed-off-by: TwoBit <robert.walenta@alumni.fh-aachen.de>